requirements.txt
numpy
pandas
matplotlib
google
google-genai
python-dotenv
Flask
google-generativeai
python==3.11.11